+++ kafka_consumer/tests/test_integration.py
@@ -5,8 +5,9 @@ import logging
 from contextlib import nullcontext as does_not_raise
 
 import pytest
-from tests.common import assert_check_kafka, metrics
+from tests.common import LEGACY_CLIENT, assert_check_kafka, metrics
 
+from datadog_checks.base import ConfigurationError
 from datadog_checks.dev.utils import get_metadata_metrics
 
 pytestmark = [pytest.mark.integration, pytest.mark.usefixtures('dd_environment')]
@@ -168,3 +169,48 @@ def test_config(
 
     assert exception_msg in caplog.text
     aggregator.assert_metrics_using_metadata(get_metadata_metrics())
+
+
+@pytest.mark.parametrize(
+    'sasl_oauth_token_provider, expected_exception',
+    [
+        pytest.param(
+            {},
+            pytest.raises(AssertionError, match="sasl_oauth_token_provider required for OAUTHBEARER sasl"),
+            id="No sasl_oauth_token_provider",
+        ),
+        pytest.param(
+            {'sasl_oauth_token_provider': {}},
+            pytest.raises(ConfigurationError, match="The `url` setting of `auth_token` reader is required"),
+            id="Empty sasl_oauth_token_provider, url missing",
+        ),
+        pytest.param(
+            {'sasl_oauth_token_provider': {'url': 'http://fake.url'}},
+            pytest.raises(ConfigurationError, match="The `client_id` setting of `auth_token` reader is required"),
+            id="client_id missing",
+        ),
+        pytest.param(
+            {'sasl_oauth_token_provider': {'url': 'http://fake.url', 'client_id': 'id'}},
+            pytest.raises(ConfigurationError, match="The `client_secret` setting of `auth_token` reader is required"),
+            id="client_secret missing",
+        ),
+        pytest.param(
+            {'sasl_oauth_token_provider': {'url': 'http://fake.url', 'client_id': 'id', 'client_secret': 'secret'}},
+            pytest.raises(Exception, match="NoBrokersAvailable"),  # Mock the expected response after library migration
+            id="valid config",
+        ),
+    ],
+)
+@pytest.mark.skipif(not LEGACY_CLIENT, reason='not implemented yet with confluent-kafka')
+def test_oauth_config(sasl_oauth_token_provider, check, expected_exception, kafka_instance):
+    kafka_instance.update(
+        {
+            'monitor_unlisted_consumer_groups': True,
+            'security_protocol': 'SASL_PLAINTEXT',
+            'sasl_mechanism': 'OAUTHBEARER',
+        }
+    )
+    kafka_instance.update(sasl_oauth_token_provider)
+
+    with expected_exception:
+        check(kafka_instance).check(kafka_instance)
